+++ apps/browser/src/background/main.background.ts
@@ -122,6 +122,8 @@ import { SyncNotifierService } from "@bitwarden/common/vault/services/sync/sync-
 import { SyncService } from "@bitwarden/common/vault/services/sync/sync.service";
 import { TotpService } from "@bitwarden/common/vault/services/totp.service";
 import {
+  IndividualVaultExportService,
+  IndividualVaultExportServiceAbstraction,
   VaultExportService,
   VaultExportServiceAbstraction,
 } from "@bitwarden/exporter/vault-export";
@@ -251,6 +253,7 @@ export default class MainBackground {
   derivedStateProvider: DerivedStateProvider;
   stateProvider: StateProvider;
   fido2Service: Fido2ServiceAbstraction;
+  individualVaultExportService: IndividualVaultExportServiceAbstraction;
 
   // Passed to the popup for Safari to workaround issues with theming, downloading, etc.
   backgroundWindow = window;
@@ -620,13 +623,18 @@ export default class MainBackground {
       this.cryptoService,
     );
 
-    this.exportService = new VaultExportService(
+    (this.individualVaultExportService = new IndividualVaultExportService(
       this.folderService,
       this.cipherService,
       this.cryptoService,
       this.cryptoFunctionService,
       this.stateService,
-    );
+    )),
+      (this.exportService = new VaultExportService(
+        this.individualVaultExportService,
+        //Setting OrgVaultExportService as null as Browser and Desktop doesn't support Organizational Export
+        null,
+      ));
     this.notificationsService = new NotificationsService(
       this.logService,
       this.syncService,
