+++ apps/web/src/app/admin-console/organizations/tools/vault-export/org-vault-export.component.ts
@@ -13,10 +13,7 @@ import { I18nService } from "@bitwarden/common/platform/abstractions/i18n.servic
 import { LogService } from "@bitwarden/common/platform/abstractions/log.service";
 import { PlatformUtilsService } from "@bitwarden/common/platform/abstractions/platform-utils.service";
 import { DialogService } from "@bitwarden/components";
-import {
-  OrgVaultExportServiceAbstraction,
-  VaultExportServiceAbstraction,
-} from "@bitwarden/exporter/vault-export";
+import { VaultExportServiceAbstraction } from "@bitwarden/exporter/vault-export";
 
 import { ExportComponent } from "../../../../tools/vault-export/export.component";
 
@@ -40,7 +37,6 @@ export class OrganizationVaultExportComponent extends ExportComponent {
     dialogService: DialogService,
     organizationService: OrganizationService,
     configService: ConfigServiceAbstraction,
-    private orgExportService: OrgVaultExportServiceAbstraction,
   ) {
     super(
       i18nService,
@@ -72,12 +68,12 @@ export class OrganizationVaultExportComponent extends ExportComponent {
 
   getExportData() {
     if (this.isFileEncryptedExport) {
-      return this.orgExportService.getPasswordProtectedExport(
-        this.filePassword,
+      return this.exportService.getOrgnizationPasswordProtectedExport(
         this.organizationId,
+        this.filePassword,
       );
     } else {
-      return this.orgExportService.getOrganizationExport(this.organizationId, this.format);
+      return this.exportService.getOrganizationExport(this.organizationId, this.format);
     }
   }
