+++ libs/exporter/src/vault-export/index.ts
@@ -2,3 +2,5 @@ export * from "./services/vault-export.service.abstraction";
 export * from "./services/vault-export.service";
 export * from "./services/org-vault-export.service.abstraction";
 export * from "./services/org-vault-export.service";
+export * from "./services/individual-vault-export.service.abstraction";
+export * from "./services/individual-vault-export.service";
