+++ libs/exporter/src/vault-export/services/individual-vault-export.service.abstraction.ts
@@ -1,0 +1,7 @@
+import { ExportFormat } from "./vault-export.service.abstraction";
+
+export abstract class IndividualVaultExportServiceAbstraction {
+  getExport: (format?: ExportFormat) => Promise<string>;
+  getPasswordProtectedExport: (password: string) => Promise<string>;
+  getFileName: (prefix?: string, extension?: string) => string;
+}
