+++ libs/exporter/src/vault-export/services/individual-vault-export.service.spec.ts
@@ -20,7 +20,7 @@ import { LoginView } from "@bitwarden/common/vault/models/view/login.view";
 
 import { BuildTestObject, GetUniqueString } from "../../../../common/spec";
 
-import { VaultExportService } from "./vault-export.service";
+import { IndividualVaultExportService } from "./individual-vault-export.service";
 
 const UserCipherViews = [
   generateCipherView(false),
@@ -139,7 +139,7 @@ function expectEqualFolders(folders: Folder[], jsonResult: string) {
 }
 
 describe("VaultExportService", () => {
-  let exportService: VaultExportService;
+  let exportService: IndividualVaultExportService;
   let cryptoFunctionService: MockProxy<CryptoFunctionService>;
   let cipherService: MockProxy<CipherService>;
   let folderService: MockProxy<FolderService>;
@@ -159,7 +159,7 @@ describe("VaultExportService", () => {
     stateService.getKdfConfig.mockResolvedValue(new KdfConfig(PBKDF2_ITERATIONS.defaultValue));
     cryptoService.encrypt.mockResolvedValue(new EncString("encrypted"));
 
-    exportService = new VaultExportService(
+    exportService = new IndividualVaultExportService(
       folderService,
       cipherService,
       cryptoService,
