+++ libs/exporter/src/vault-export/services/org-vault-export.service.abstraction.ts
@@ -1,6 +1,6 @@
 import { ExportFormat } from "./vault-export.service.abstraction";
 
-export abstract class OrgVaultExportServiceAbstraction {
+export abstract class OrganizationVaultExportServiceAbstraction {
   getPasswordProtectedExport: (password: string, organizationId?: string) => Promise<string>;
   getOrganizationExport: (organizationId: string, format?: ExportFormat) => Promise<string>;
   getFileName: (prefix?: string, extension?: string) => string;
